// src/entities/auction/model/types.ts
export type AuctionStatus =
  | "Planning"
  | "Auction"
  | "DeterminateWinner"
  | "WaitDeal"
  | "InProgress"
  | "Finished"
  | "Stopped"
  | "Canceled"
  | "Unknown";

export type AuctionType = "Request" | "Up" | "Down" | "FixPrice" | "Unknown";

export type TradingStatus =
  | "NotParticipating"
  | "Leading"
  | "Losing"
  | "OnPending"
  | "Confirmed"
  | "ChoosingWinner"
  | "Winner"
  | "Accepted"
  | "Unknown";

export type BidMeasurementType = "PerRoute" | "PerKm" | "Unknown";
export type OperationType = "Loading" | "Unloading" | "Unknown";
export type PaymentDelayType = "CalendarDays" | "WorkDays" | "Unknown";

export interface AuctionListItemMain {
  id: number;
  cargo_num: string;
  cargo_date: string;
  auc_type: AuctionType;
  order_uid: string;
  created_at: string;
  priority_sort: number;
  is_assembly: boolean;
  price_per_km: number | null;
}

export interface AuctionListItemOrganizer {
  subscriber_id: number;
  organization_id: number;
  organization_name: string;
  organization_inn: string;
  organization_kpp: string;
  is_hide_organization: boolean;
}

export interface AuctionListItemRoutePoint {
  city: string;
  address: string;
  date: string;
  city_gc_id: number;
  points_count: number;
}

export interface AuctionListItemRoute {
  load: AuctionListItemRoutePoint;
  unload: AuctionListItemRoutePoint;
}

export interface LoadingTypes {
  side: boolean;
  top: boolean;
  rear: boolean;
  full: boolean;
}

export interface Docs {
  tir: boolean;
  cmr: boolean;
  t1: boolean;
  med: boolean;
}

export interface CarRequirements {
  type: string;
  weight: number | null;
  volume: number | null;
  width: number | null;
  length: number | null;
  height: number | null;
}

export interface AuctionListItemCargo {
  name: string;
  weight: number;
  volume: number;
  body_type: string;
  truck_count: number;
  is_cargo: boolean;
  is_international: boolean | null;
  containered: boolean | null;
  incoterms: string | null;
  conics: number | null;
  belts: number | null;
  adr: number | null;
  coupling: boolean | null;
  air_pass: boolean | null;
  low_loader: boolean | null;
  additional_load: boolean | null;
  temp_from: number | null;
  temp_to: number | null;
  loading_types: LoadingTypes;
  docs: Docs;
  car: CarRequirements | null;
}

export interface AuctionListItemTradingPrice {
  start: number;
  current: number;
  current_no_vat: number;
}

export interface AuctionListItemTradingYour {
  bet: boolean;
  last_bet: number | null;
}

export interface AuctionListItemTrading {
  status: AuctionStatus;
  status_mobile: Exclude<
    TradingStatus,
    "OnPending" | "ChoosingWinner" | "Accepted"
  >;
  start_time: string;
  stop_time: string;
  bid_measurement_type: BidMeasurementType | null;
  can_set_bet: boolean;
  allow_counter_bets: boolean;
  hide_points_address_and_contacts: boolean;
  direction: string | null;
  comment: string | null;
  is_bidder: boolean;
  is_available: boolean;
  is_accredited: boolean;
  is_favorite: boolean;
  price: AuctionListItemTradingPrice | null;
  your: AuctionListItemTradingYour | null;
  red_bet_with_vat: boolean;
  red_bet_no_vat: boolean;
  is_last_bet_with_vat: boolean | null;
}

export interface AuctionListItemPayment {
  form: string;
  currency_code: string;
  consignor: string | null;
  consignee: string | null;
}

export interface AuctionListItem {
  main: AuctionListItemMain;
  organizer: AuctionListItemOrganizer;
  route: AuctionListItemRoute;
  cargo: AuctionListItemCargo;
  trading: AuctionListItemTrading;
  payment: AuctionListItemPayment;
}